+++ arduino/random_numbers.go
@@ -5,10 +5,10 @@ package arduino
 //
 //go:wasm-module arduino
 //export randomArduino
-func Random() int64
+func Random() int32
 
 // Equivalent to Arduino's randomSeed https://www.arduino.cc/reference/en/language/functions/random-numbers/randomseed/
 //
 //go:wasm-module arduino
 //export randomSeed
-func RandomSeed(seed int64)
+func RandomSeed(seed int32)
